+++ release_notes.md
@@ -6,6 +6,7 @@ PREESM Changelog
 
 ### New Feature
 * Added a new tab in the scenario to include power and energy information for the architecture and for the algorithm being executed on the different resources. The possibility to include a performance objective is also added;
+* Add support for exporting hierarchical PiSDF graphs;
 
 ### Changes
 * Fix gantt viewer PE order;
@@ -13,6 +14,7 @@ PREESM Changelog
 * Fix Slam parser;
 * MPPA2Explicit now is called MPPA2Cluster;
 * MPPA2IOExplicit now is called MPPA2IO;
+* Update PiSDF consistency checker;
 
 ### Bug fix
 * Fix #166: forces scheduler to put init and end actors on same core;